Fix casting bug

+++ lobcder-worker/src/main/java/nl/uva/cs/lobcder/Util.java
@@ -94,7 +94,7 @@ class Util {
     }
 
     static double getProgressThresshold() throws IOException {
-        return Integer.valueOf(getProperties().getProperty(("progress.thresshold"), "10"));
+        return Double.valueOf(getProperties().getProperty(("progress.thresshold"), "10"));
     }
 
     static double getProgressThressholdCoefficient() throws IOException {
